API: libpe_status,libpe_rules,libcrmcommon: move enum action_tasks to libcrmcommon
