+++ include/crm/common/scheduler.h
@@ -10,6 +10,7 @@
 #ifndef PCMK__CRM_COMMON_SCHEDULER__H
 #  define PCMK__CRM_COMMON_SCHEDULER__H
 
+#include <crm/common/actions.h>
 #include <crm/common/resources.h>
 #include <crm/common/roles.h>
